requirements.txt
python-telegram-bot==20.7
openai==1.3.7
python-dotenv==1.0.0
pydub==0.25.1
speechrecognition==3.10.0
requests==2.31.0
aiohttp==3.9.1
matplotlib==3.7.2
numpy==1.24.3
